HBASE-25005 Refactor CatalogJanitor (#2373)
