+++ hbase-server/src/test/java/org/apache/hadoop/hbase/master/janitor/TestCatalogJanitorCluster.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.hadoop.hbase.master;
+package org.apache.hadoop.hbase.master.janitor;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -50,12 +50,13 @@ import org.junit.rules.TestName;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-@Category({MasterTests.class, LargeTests.class})
+@Category({ MasterTests.class, LargeTests.class })
 public class TestCatalogJanitorCluster {
   private static final Logger LOG = LoggerFactory.getLogger(TestCatalogJanitorCluster.class);
+
   @ClassRule
   public static final HBaseClassTestRule CLASS_RULE =
-      HBaseClassTestRule.forClass(TestCatalogJanitorCluster.class);
+    HBaseClassTestRule.forClass(TestCatalogJanitorCluster.class);
 
   @Rule
   public final TestName name = new TestName();
@@ -70,24 +71,16 @@ public class TestCatalogJanitorCluster {
   @Before
   public void before() throws Exception {
     TEST_UTIL.startMiniCluster();
-    TEST_UTIL.createMultiRegionTable(T1, new byte [][] {HConstants.CATALOG_FAMILY});
-    TEST_UTIL.createMultiRegionTable(T2, new byte [][] {HConstants.CATALOG_FAMILY});
-    TEST_UTIL.createMultiRegionTable(T3, new byte [][] {HConstants.CATALOG_FAMILY});
+    TEST_UTIL.createMultiRegionTable(T1, new byte[][] { HConstants.CATALOG_FAMILY });
+    TEST_UTIL.createMultiRegionTable(T2, new byte[][] { HConstants.CATALOG_FAMILY });
+    TEST_UTIL.createMultiRegionTable(T3, new byte[][] { HConstants.CATALOG_FAMILY });
 
-    final byte[][] keysForT4 = {
-      Bytes.toBytes("aa"),
-      Bytes.toBytes("bb"),
-      Bytes.toBytes("cc"),
-      Bytes.toBytes("dd")
-    };
+    final byte[][] keysForT4 =
+      { Bytes.toBytes("aa"), Bytes.toBytes("bb"), Bytes.toBytes("cc"), Bytes.toBytes("dd") };
 
     TEST_UTIL.createTable(T4, HConstants.CATALOG_FAMILY, keysForT4);
 
-    final byte[][] keysForT5 = {
-      Bytes.toBytes("bb"),
-      Bytes.toBytes("cc"),
-      Bytes.toBytes("dd")
-    };
+    final byte[][] keysForT5 = { Bytes.toBytes("bb"), Bytes.toBytes("cc"), Bytes.toBytes("dd") };
 
     TEST_UTIL.createTable(T5, HConstants.CATALOG_FAMILY, keysForT5);
   }
@@ -98,11 +91,10 @@ public class TestCatalogJanitorCluster {
   }
 
   /**
-   * Fat method where we start with a fat hbase:meta and then gradually intro
-   * problems running catalogjanitor for each to ensure it triggers complaint.
-   * Do one big method because takes a while to build up the context we need.
-   * We create three tables and then make holes, overlaps, add unknown servers
-   * and empty out regioninfo columns. Each should up counts in the
+   * Fat method where we start with a fat hbase:meta and then gradually intro problems running
+   * catalogjanitor for each to ensure it triggers complaint. Do one big method because takes a
+   * while to build up the context we need. We create three tables and then make holes, overlaps,
+   * add unknown servers and empty out regioninfo columns. Each should up counts in the
    * CatalogJanitor.Report produced.
    */
   @Test
@@ -110,34 +102,33 @@ public class TestCatalogJanitorCluster {
     CatalogJanitor janitor = TEST_UTIL.getHBaseCluster().getMaster().getCatalogJanitor();
     RegionStateStore regionStateStore =
       TEST_UTIL.getHBaseCluster().getMaster().getAssignmentManager().getRegionStateStore();
-    int gc = janitor.scan();
-    CatalogJanitor.Report report = janitor.getLastReport();
+    janitor.scan();
+    Report report = janitor.getLastReport();
     // Assert no problems.
     assertTrue(report.isEmpty());
     // Now remove first region in table t2 to see if catalogjanitor scan notices.
     List<RegionInfo> t2Ris = MetaTableAccessor.getTableRegions(TEST_UTIL.getConnection(), T2);
     regionStateStore.deleteRegion(t2Ris.get(0));
-    gc = janitor.scan();
+    janitor.scan();
     report = janitor.getLastReport();
     assertFalse(report.isEmpty());
     assertEquals(1, report.getHoles().size());
-    assertTrue(
-      report.getHoles().get(0).getFirst().getTable().equals(RegionInfo.UNDEFINED.getTable()));
+    assertTrue(report.getHoles().get(0).getFirst().getTable()
+      .equals(RegionInfoBuilder.UNDEFINED.getTable()));
     assertTrue(report.getHoles().get(0).getSecond().getTable().equals(T2));
     assertEquals(0, report.getOverlaps().size());
     // Next, add overlaps to first row in t3
     List<RegionInfo> t3Ris = MetaTableAccessor.getTableRegions(TEST_UTIL.getConnection(), T3);
     RegionInfo ri = t3Ris.get(0);
-    RegionInfo newRi1 = RegionInfoBuilder.newBuilder(ri.getTable()).
-        setStartKey(incrementRow(ri.getStartKey())).
-        setEndKey(incrementRow(ri.getEndKey())).build();
+    RegionInfo newRi1 = RegionInfoBuilder.newBuilder(ri.getTable())
+      .setStartKey(incrementRow(ri.getStartKey())).setEndKey(incrementRow(ri.getEndKey())).build();
     Put p1 = MetaTableAccessor.makePutFromRegionInfo(newRi1, System.currentTimeMillis());
-    RegionInfo newRi2 = RegionInfoBuilder.newBuilder(newRi1.getTable()).
-        setStartKey(incrementRow(newRi1.getStartKey())).
-        setEndKey(incrementRow(newRi1.getEndKey())).build();
+    RegionInfo newRi2 = RegionInfoBuilder.newBuilder(newRi1.getTable())
+      .setStartKey(incrementRow(newRi1.getStartKey())).setEndKey(incrementRow(newRi1.getEndKey()))
+      .build();
     Put p2 = MetaTableAccessor.makePutFromRegionInfo(newRi2, System.currentTimeMillis());
     MetaTableAccessor.putsToMetaTable(TEST_UTIL.getConnection(), Arrays.asList(p1, p2));
-    gc = janitor.scan();
+    janitor.scan();
     report = janitor.getLastReport();
     assertFalse(report.isEmpty());
     // We added two overlaps so total three.
@@ -151,10 +142,10 @@ public class TestCatalogJanitorCluster {
     List<RegionInfo> t1Ris = MetaTableAccessor.getTableRegions(TEST_UTIL.getConnection(), T1);
     RegionInfo t1Ri1 = t1Ris.get(1);
     Put pServer = new Put(t1Ri1.getRegionName());
-    pServer.addColumn(HConstants.CATALOG_FAMILY,
-        CatalogFamilyFormat.getServerColumn(0), Bytes.toBytes("bad.server.example.org:1234"));
+    pServer.addColumn(HConstants.CATALOG_FAMILY, CatalogFamilyFormat.getServerColumn(0),
+      Bytes.toBytes("bad.server.example.org:1234"));
     MetaTableAccessor.putsToMetaTable(TEST_UTIL.getConnection(), Arrays.asList(pServer));
-    gc = janitor.scan();
+    janitor.scan();
     report = janitor.getLastReport();
     assertFalse(report.isEmpty());
     assertEquals(1, report.getUnknownServers().size());
@@ -163,8 +154,8 @@ public class TestCatalogJanitorCluster {
     // break if this happens.
     LOG.info("Make null info:server");
     Put emptyInfoServerPut = new Put(t1Ri1.getRegionName());
-    emptyInfoServerPut.addColumn(HConstants.CATALOG_FAMILY,
-        CatalogFamilyFormat.getServerColumn(0), Bytes.toBytes(""));
+    emptyInfoServerPut.addColumn(HConstants.CATALOG_FAMILY, CatalogFamilyFormat.getServerColumn(0),
+      Bytes.toBytes(""));
     MetaTableAccessor.putsToMetaTable(TEST_UTIL.getConnection(), Arrays.asList(emptyInfoServerPut));
     janitor.scan();
     report = janitor.getLastReport();
@@ -172,8 +163,8 @@ public class TestCatalogJanitorCluster {
     // Mke an empty regioninfo in t1.
     RegionInfo t1Ri2 = t1Ris.get(2);
     Put pEmptyRI = new Put(t1Ri2.getRegionName());
-    pEmptyRI.addColumn(HConstants.CATALOG_FAMILY,
-        HConstants.REGIONINFO_QUALIFIER, HConstants.EMPTY_BYTE_ARRAY);
+    pEmptyRI.addColumn(HConstants.CATALOG_FAMILY, HConstants.REGIONINFO_QUALIFIER,
+      HConstants.EMPTY_BYTE_ARRAY);
     MetaTableAccessor.putsToMetaTable(TEST_UTIL.getConnection(), Arrays.asList(pEmptyRI));
     janitor.scan();
     report = janitor.getLastReport();
@@ -183,16 +174,15 @@ public class TestCatalogJanitorCluster {
     int overlapsReported = report.getOverlaps().size();
 
     // Test the case for T4
-    //    r1: [aa, bb), r2: [cc, dd), r3: [a, cc)
+    // r1: [aa, bb), r2: [cc, dd), r3: [a, cc)
     // Make sure only overlaps and no holes are reported.
     List<RegionInfo> t4Ris = MetaTableAccessor.getTableRegions(TEST_UTIL.getConnection(), T4);
     // delete the region [bb, cc)
     regionStateStore.deleteRegion(t4Ris.get(2));
 
     // add a new region [a, cc)
-    RegionInfo newRiT4 = RegionInfoBuilder.newBuilder(T4).
-      setStartKey("a".getBytes()).
-      setEndKey("cc".getBytes()).build();
+    RegionInfo newRiT4 = RegionInfoBuilder.newBuilder(T4).setStartKey("a".getBytes())
+      .setEndKey("cc".getBytes()).build();
     Put putForT4 = MetaTableAccessor.makePutFromRegionInfo(newRiT4, System.currentTimeMillis());
     MetaTableAccessor.putsToMetaTable(TEST_UTIL.getConnection(), Arrays.asList(putForT4));
 
@@ -206,16 +196,15 @@ public class TestCatalogJanitorCluster {
     overlapsReported = report.getOverlaps().size();
 
     // Test the case for T5
-    //    r0: [, bb), r1: [a, g), r2: [bb, cc), r3: [dd, )
+    // r0: [, bb), r1: [a, g), r2: [bb, cc), r3: [dd, )
     // Make sure only overlaps and no holes are reported.
     List<RegionInfo> t5Ris = MetaTableAccessor.getTableRegions(TEST_UTIL.getConnection(), T5);
     // delete the region [cc, dd)
     regionStateStore.deleteRegion(t5Ris.get(2));
 
     // add a new region [a, g)
-    RegionInfo newRiT5 = RegionInfoBuilder.newBuilder(T5).
-      setStartKey("a".getBytes()).
-      setEndKey("g".getBytes()).build();
+    RegionInfo newRiT5 = RegionInfoBuilder.newBuilder(T5).setStartKey("a".getBytes())
+      .setEndKey("g".getBytes()).build();
     Put putForT5 = MetaTableAccessor.makePutFromRegionInfo(newRiT5, System.currentTimeMillis());
     MetaTableAccessor.putsToMetaTable(TEST_UTIL.getConnection(), Arrays.asList(putForT5));
 
@@ -230,11 +219,11 @@ public class TestCatalogJanitorCluster {
   /**
    * Take last byte and add one to it.
    */
-  private static byte [] incrementRow(byte [] row) {
+  private static byte[] incrementRow(byte[] row) {
     if (row.length == 0) {
-      return new byte []{'0'};
+      return new byte[] { '0' };
     }
-    row[row.length - 1] = (byte)(((int)row[row.length - 1]) + 1);
+    row[row.length - 1] = (byte) (((int) row[row.length - 1]) + 1);
     return row;
   }
 
@@ -242,23 +231,23 @@ public class TestCatalogJanitorCluster {
   public void testHoles() throws IOException {
     CatalogJanitor janitor = TEST_UTIL.getHBaseCluster().getMaster().getCatalogJanitor();
 
-    CatalogJanitor.Report report = janitor.getLastReport();
+    Report report = janitor.getLastReport();
     // Assert no problems.
     assertTrue(report.isEmpty());
-    //Verify start and end region holes
+    // Verify start and end region holes
     verifyCornerHoles(janitor, T1);
-    //Verify start and end region holes
+    // Verify start and end region holes
     verifyCornerHoles(janitor, T2);
     verifyMiddleHole(janitor);
-    //Verify that MetaFixer is able to fix these holes
+    // Verify that MetaFixer is able to fix these holes
     fixHoles(janitor);
   }
 
   private void fixHoles(CatalogJanitor janitor) throws IOException {
     MetaFixer metaFixer = new MetaFixer(TEST_UTIL.getHBaseCluster().getMaster());
     janitor.scan();
-    CatalogJanitor.Report report = janitor.getLastReport();
-    //Verify total number of holes, 2 in t1 and t2 each and one in t3
+    Report report = janitor.getLastReport();
+    // Verify total number of holes, 2 in t1 and t2 each and one in t3
     assertEquals("Number of holes are not matching", 5, report.getHoles().size());
     metaFixer.fix();
     janitor.scan();
@@ -267,7 +256,7 @@ public class TestCatalogJanitorCluster {
   }
 
   private void verifyMiddleHole(CatalogJanitor janitor) throws IOException {
-    //Verify middle holes
+    // Verify middle holes
     RegionInfo firstRegion = getRegionInfo(T3, "".getBytes());
     RegionInfo secondRegion = getRegionInfo(T3, "bbb".getBytes());
     RegionInfo thirdRegion = getRegionInfo(T3, "ccc".getBytes());
@@ -293,8 +282,8 @@ public class TestCatalogJanitorCluster {
 
     assertEquals(1, holes.size());
     Pair<RegionInfo, RegionInfo> regionInfoRegionInfoPair = holes.get(0);
-    assertTrue(
-      regionInfoRegionInfoPair.getFirst().getTable().equals(RegionInfo.UNDEFINED.getTable()));
+    assertTrue(regionInfoRegionInfoPair.getFirst().getTable()
+      .equals(RegionInfoBuilder.UNDEFINED.getTable()));
     assertTrue(regionInfoRegionInfoPair.getSecond().getTable().equals(tableName));
     assertTrue(
       regionInfoRegionInfoPair.getSecond().getEncodedName().equals(secondRegion.getEncodedName()));
@@ -307,20 +296,20 @@ public class TestCatalogJanitorCluster {
     regionInfoRegionInfoPair = holes.get(1);
     assertTrue(regionInfoRegionInfoPair.getFirst().getEncodedName()
       .equals(secondLastRegion.getEncodedName()));
-    assertTrue(
-      regionInfoRegionInfoPair.getSecond().getTable().equals(RegionInfo.UNDEFINED.getTable()));
+    assertTrue(regionInfoRegionInfoPair.getSecond().getTable()
+      .equals(RegionInfoBuilder.UNDEFINED.getTable()));
   }
 
-  //Get Holes filter by table
+  // Get Holes filter by table
   private LinkedList<Pair<RegionInfo, RegionInfo>> getHoles(CatalogJanitor janitor,
     TableName tableName) throws IOException {
     janitor.scan();
-    CatalogJanitor.Report lastReport = janitor.getLastReport();
+    Report lastReport = janitor.getLastReport();
     assertFalse(lastReport.isEmpty());
     LinkedList<Pair<RegionInfo, RegionInfo>> holes = new LinkedList<>();
     for (Pair<RegionInfo, RegionInfo> hole : lastReport.getHoles()) {
-      if (hole.getFirst().getTable().equals(tableName) || hole.getSecond().getTable()
-        .equals(tableName)) {
+      if (hole.getFirst().getTable().equals(tableName) ||
+        hole.getSecond().getTable().equals(tableName)) {
         holes.add(hole);
       }
     }
