+++ hbase-server/src/test/java/org/apache/hadoop/hbase/master/janitor/TestCatalogJanitorInMemoryStates.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.hadoop.hbase.master;
+package org.apache.hadoop.hbase.master.janitor;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
@@ -25,7 +25,6 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
-import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hbase.CatalogFamilyFormat;
 import org.apache.hadoop.hbase.HBaseClassTestRule;
 import org.apache.hadoop.hbase.HBaseTestingUtility;
@@ -33,6 +32,7 @@ import org.apache.hadoop.hbase.HRegionLocation;
 import org.apache.hadoop.hbase.MetaMockingUtil;
 import org.apache.hadoop.hbase.MetaTableAccessor;
 import org.apache.hadoop.hbase.TableName;
+import org.apache.hadoop.hbase.TableNameTestRule;
 import org.apache.hadoop.hbase.client.Admin;
 import org.apache.hadoop.hbase.client.Connection;
 import org.apache.hadoop.hbase.client.ConnectionFactory;
@@ -41,6 +41,8 @@ import org.apache.hadoop.hbase.client.RegionInfo;
 import org.apache.hadoop.hbase.client.RegionLocator;
 import org.apache.hadoop.hbase.client.Result;
 import org.apache.hadoop.hbase.client.Table;
+import org.apache.hadoop.hbase.master.HMaster;
+import org.apache.hadoop.hbase.master.ServerManager;
 import org.apache.hadoop.hbase.master.assignment.AssignmentManager;
 import org.apache.hadoop.hbase.testclassification.MasterTests;
 import org.apache.hadoop.hbase.testclassification.MediumTests;
@@ -53,37 +55,30 @@ import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
-import org.junit.rules.TestName;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-@Category({MasterTests.class, MediumTests.class})
+@Category({ MasterTests.class, MediumTests.class })
 public class TestCatalogJanitorInMemoryStates {
 
   @ClassRule
   public static final HBaseClassTestRule CLASS_RULE =
-      HBaseClassTestRule.forClass(TestCatalogJanitorInMemoryStates.class);
+    HBaseClassTestRule.forClass(TestCatalogJanitorInMemoryStates.class);
 
   private static final Logger LOG = LoggerFactory.getLogger(TestCatalogJanitorInMemoryStates.class);
-  @Rule public final TestName name = new TestName();
-  protected final static HBaseTestingUtility TEST_UTIL = new HBaseTestingUtility();
-  private static byte [] ROW = Bytes.toBytes("testRow");
-  private static byte [] FAMILY = Bytes.toBytes("testFamily");
-  private static byte [] QUALIFIER = Bytes.toBytes("testQualifier");
-  private static byte [] VALUE = Bytes.toBytes("testValue");
 
-  /**
-   * @throws java.lang.Exception
-   */
+  private static final HBaseTestingUtility TEST_UTIL = new HBaseTestingUtility();
+
+  private static byte[] FAMILY = Bytes.toBytes("testFamily");
+
+  @Rule
+  public final TableNameTestRule name = new TableNameTestRule();
+
   @BeforeClass
   public static void setUpBeforeClass() throws Exception {
-    Configuration conf = TEST_UTIL.getConfiguration();
     TEST_UTIL.startMiniCluster(1);
   }
 
-  /**
-   * @throws java.lang.Exception
-   */
   @AfterClass
   public static void tearDownAfterClass() throws Exception {
     TEST_UTIL.shutdownMiniCluster();
@@ -94,17 +89,17 @@ public class TestCatalogJanitorInMemoryStates {
    */
   @Test
   public void testInMemoryParentCleanup()
-      throws IOException, InterruptedException, ExecutionException {
-    final AssignmentManager am = TEST_UTIL.getHBaseCluster().getMaster().getAssignmentManager();
-    final ServerManager sm = TEST_UTIL.getHBaseCluster().getMaster().getServerManager();
-    final CatalogJanitor janitor = TEST_UTIL.getHBaseCluster().getMaster().getCatalogJanitor();
+    throws IOException, InterruptedException, ExecutionException {
+    HMaster master = TEST_UTIL.getHBaseCluster().getMaster();
+    final AssignmentManager am = master.getAssignmentManager();
+    final ServerManager sm = master.getServerManager();
 
     Admin admin = TEST_UTIL.getAdmin();
     admin.catalogJanitorSwitch(false);
 
-    final TableName tableName = TableName.valueOf(name.getMethodName());
+    final TableName tableName = name.getTableName();
     Table t = TEST_UTIL.createTable(tableName, FAMILY);
-    int rowCount = TEST_UTIL.loadTable(t, FAMILY, false);
+    TEST_UTIL.loadTable(t, FAMILY, false);
 
     RegionLocator locator = TEST_UTIL.getConnection().getRegionLocator(tableName);
     List<HRegionLocation> allRegionLocations = locator.getAllRegionLocations();
@@ -117,18 +112,18 @@ public class TestCatalogJanitorInMemoryStates {
     assertNotNull("Should have found daughter regions for " + parent, daughters);
 
     assertTrue("Parent region should exist in RegionStates",
-        am.getRegionStates().isRegionInRegionStates(parent.getRegion()));
+      am.getRegionStates().isRegionInRegionStates(parent.getRegion()));
     assertTrue("Parent region should exist in ServerManager",
-        sm.isRegionInServerManagerStates(parent.getRegion()));
+      sm.isRegionInServerManagerStates(parent.getRegion()));
 
     // clean the parent
     Result r = MetaMockingUtil.getMetaTableRowResult(parent.getRegion(), null,
-        daughters.get(0).getRegion(), daughters.get(1).getRegion());
-    janitor.cleanParent(parent.getRegion(), r);
+      daughters.get(0).getRegion(), daughters.get(1).getRegion());
+    CatalogJanitor.cleanParent(master, parent.getRegion(), r);
     assertFalse("Parent region should have been removed from RegionStates",
-        am.getRegionStates().isRegionInRegionStates(parent.getRegion()));
+      am.getRegionStates().isRegionInRegionStates(parent.getRegion()));
     assertFalse("Parent region should have been removed from ServerManager",
-        sm.isRegionInServerManagerStates(parent.getRegion()));
+      sm.isRegionInServerManagerStates(parent.getRegion()));
 
   }
 
@@ -138,7 +133,7 @@ public class TestCatalogJanitorInMemoryStates {
    * @return List of region locations
    */
   private List<HRegionLocation> splitRegion(final RegionInfo r)
-      throws IOException, InterruptedException, ExecutionException {
+    throws IOException, InterruptedException, ExecutionException {
     List<HRegionLocation> locations = new ArrayList<>();
     // Split this table in two.
     Admin admin = TEST_UTIL.getAdmin();
@@ -157,17 +152,16 @@ public class TestCatalogJanitorInMemoryStates {
   }
 
   /*
-   * Wait on region split. May return because we waited long enough on the split
-   * and it didn't happen.  Caller should check.
+   * Wait on region split. May return because we waited long enough on the split and it didn't
+   * happen. Caller should check.
    * @param r
    * @return Daughter regions; caller needs to check table actually split.
    */
-  private PairOfSameType<RegionInfo> waitOnDaughters(final RegionInfo r)
-      throws IOException {
+  private PairOfSameType<RegionInfo> waitOnDaughters(final RegionInfo r) throws IOException {
     long start = System.currentTimeMillis();
     PairOfSameType<RegionInfo> pair = null;
     try (Connection conn = ConnectionFactory.createConnection(TEST_UTIL.getConfiguration());
-         Table metaTable = conn.getTable(TableName.META_TABLE_NAME)) {
+      Table metaTable = conn.getTable(TableName.META_TABLE_NAME)) {
       Result result = null;
       RegionInfo region = null;
       while ((System.currentTimeMillis() - start) < 60000) {
