+++ hbase-server/src/test/java/org/apache/hadoop/hbase/master/janitor/TestMetaFixerNoCluster.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.hadoop.hbase.master;
+package org.apache.hadoop.hbase.master.janitor;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -24,7 +24,6 @@ import static org.junit.Assert.assertTrue;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.SortedSet;
-
 import org.apache.hadoop.hbase.HBaseClassTestRule;
 import org.apache.hadoop.hbase.TableName;
 import org.apache.hadoop.hbase.client.RegionInfo;
@@ -33,48 +32,39 @@ import org.apache.hadoop.hbase.testclassification.MasterTests;
 import org.apache.hadoop.hbase.testclassification.SmallTests;
 import org.apache.hadoop.hbase.util.Bytes;
 import org.apache.hadoop.hbase.util.Pair;
-
 import org.junit.ClassRule;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 
-
-
 /**
- * Test small utility methods inside {@link MetaFixer}.
- * For cluster tests see {@link TestMetaFixer}
+ * Test small utility methods inside {@link MetaFixer}. For cluster tests see {@link TestMetaFixer}
  */
-@Category({MasterTests.class, SmallTests.class})
+@Category({ MasterTests.class, SmallTests.class })
 public class TestMetaFixerNoCluster {
   @ClassRule
   public static final HBaseClassTestRule CLASS_RULE =
-      HBaseClassTestRule.forClass(TestMetaFixerNoCluster.class);
-  private static byte [] A = Bytes.toBytes("a");
-  private static byte [] B = Bytes.toBytes("b");
-  private static byte [] C = Bytes.toBytes("c");
-  private static byte [] D = Bytes.toBytes("d");
+    HBaseClassTestRule.forClass(TestMetaFixerNoCluster.class);
+  private static byte[] A = Bytes.toBytes("a");
+  private static byte[] B = Bytes.toBytes("b");
+  private static byte[] C = Bytes.toBytes("c");
+  private static byte[] D = Bytes.toBytes("d");
   private static RegionInfo ALL = RegionInfoBuilder.FIRST_META_REGIONINFO;
-  private static RegionInfo _ARI = RegionInfoBuilder.newBuilder(TableName.META_TABLE_NAME).
-      setEndKey(A).build();
-  private static RegionInfo _BRI = RegionInfoBuilder.newBuilder(TableName.META_TABLE_NAME).
-          setEndKey(B).build();
-  private static RegionInfo ABRI = RegionInfoBuilder.newBuilder(TableName.META_TABLE_NAME).
-          setStartKey(A).setEndKey(B).build();
-  private static RegionInfo ACRI =
-      org.apache.hadoop.hbase.client.RegionInfoBuilder.newBuilder(TableName.META_TABLE_NAME).
-          setStartKey(A).setEndKey(C).build();
-  private static RegionInfo CDRI =
-      org.apache.hadoop.hbase.client.RegionInfoBuilder.newBuilder(TableName.META_TABLE_NAME).
-          setStartKey(C).setEndKey(D).build();
-  private static RegionInfo ADRI =
-      org.apache.hadoop.hbase.client.RegionInfoBuilder.newBuilder(TableName.META_TABLE_NAME).
-          setStartKey(A).setEndKey(D).build();
-  private static RegionInfo D_RI =
-      org.apache.hadoop.hbase.client.RegionInfoBuilder.newBuilder(TableName.META_TABLE_NAME).
-          setStartKey(D).build();
-  private static RegionInfo C_RI =
-      org.apache.hadoop.hbase.client.RegionInfoBuilder.newBuilder(TableName.META_TABLE_NAME).
-          setStartKey(C).build();
+  private static RegionInfo _ARI =
+    RegionInfoBuilder.newBuilder(TableName.META_TABLE_NAME).setEndKey(A).build();
+  private static RegionInfo _BRI =
+    RegionInfoBuilder.newBuilder(TableName.META_TABLE_NAME).setEndKey(B).build();
+  private static RegionInfo ABRI =
+    RegionInfoBuilder.newBuilder(TableName.META_TABLE_NAME).setStartKey(A).setEndKey(B).build();
+  private static RegionInfo ACRI = org.apache.hadoop.hbase.client.RegionInfoBuilder
+    .newBuilder(TableName.META_TABLE_NAME).setStartKey(A).setEndKey(C).build();
+  private static RegionInfo CDRI = org.apache.hadoop.hbase.client.RegionInfoBuilder
+    .newBuilder(TableName.META_TABLE_NAME).setStartKey(C).setEndKey(D).build();
+  private static RegionInfo ADRI = org.apache.hadoop.hbase.client.RegionInfoBuilder
+    .newBuilder(TableName.META_TABLE_NAME).setStartKey(A).setEndKey(D).build();
+  private static RegionInfo D_RI = org.apache.hadoop.hbase.client.RegionInfoBuilder
+    .newBuilder(TableName.META_TABLE_NAME).setStartKey(D).build();
+  private static RegionInfo C_RI = org.apache.hadoop.hbase.client.RegionInfoBuilder
+    .newBuilder(TableName.META_TABLE_NAME).setStartKey(C).build();
 
   @Test
   public void testGetRegionInfoWithLargestEndKey() {
